+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [DAV-2698] Add SDK package and built-in authentication
 - [DAV-2359] Add wallet context support
+- [DAV-3597] Fix authentication issue in Safari
 
 ### v1.4.0